backend/requirements.txt
Flask
python-dotenv
requests
solana
fuzzywuzzy
python-Levenshtein
scikit-learn
firebase-admin
